+++ src/main/java/ru/stepup/model/ProductRegisterType.java
@@ -15,7 +15,7 @@ public class ProductRegisterType {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "internal_id")
-    private Integer internalId;
+    private Long internalId;
 
     @Column(name = "value", unique = true, nullable = false)
     private String value;
