+++ src/main/java/ru/stepup/repository/AgreementRepository.java
@@ -3,6 +3,6 @@ package ru.stepup.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.stepup.model.Agreement;
 
-public interface AgreementRepository extends JpaRepository<Agreement, Integer> {
+public interface AgreementRepository extends JpaRepository<Agreement, Long> {
     Agreement findByNumber(String number);
 }
