+++ src/main/java/ru/stepup/repository/ProductClassRepository.java
@@ -3,6 +3,6 @@ package ru.stepup.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.stepup.model.ProductClass;
 
-public interface ProductClassRepository extends JpaRepository<ProductClass, Integer> {
+public interface ProductClassRepository extends JpaRepository<ProductClass, Long> {
     ProductClass findByValue(String value);
 }
