+++ src/main/java/ru/stepup/repository/ProductRegisterRepository.java
@@ -3,6 +3,6 @@ package ru.stepup.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.stepup.model.ProductRegister;
 
-public interface ProductRegisterRepository extends JpaRepository<ProductRegister, Integer> {
-    ProductRegister findByProductIdAndProductRegisterType_Value(Integer productId, String productRegisterTypeValue);
+public interface ProductRegisterRepository extends JpaRepository<ProductRegister, Long> {
+    ProductRegister findByProductIdAndProductRegisterType_Value(Long productId, String productRegisterTypeValue);
 }
